Move load_lc_mock_chunk function to load_lc_mock.py
